refactor: backend code cleanup
