+++ tests/health.test.ts
@@ -8,6 +8,7 @@ import {
 import { HeraldHttpApi, HttpHealthLive, HttpS3Live } from "../src/Http.ts";
 import { HeraldConfig } from "../src/Config/Layer.ts";
 import { S3ClientLive } from "../src/Backends/S3/Client.ts";
+import { SwiftClientLive } from "../src/Backends/Swift/Client.ts";
 import { S3XmlLive } from "../src/Services/S3Xml.ts";
 import { BackendResolverLive } from "../src/Services/BackendResolver.ts";
 import { EffectAssert, testEffect } from "./utils.ts";
@@ -22,8 +23,9 @@ testEffect("health/getStatus", () =>
     const ApiWithRequirements = HttpApiBuilder.api(HeraldHttpApi).pipe(
       Layer.provide(HttpHealthLive),
       Layer.provide(HttpS3Live),
-      Layer.provide(S3ClientLive),
       Layer.provide(BackendResolverLive),
+      Layer.provide(S3ClientLive),
+      Layer.provide(SwiftClientLive),
       Layer.provide(S3XmlLive),
       Layer.provide(HeraldConfigLive),
       Layer.provide(FetchHttpClient.layer),
